Want way to check for API_VERSION update (#161)
