+++ opte-api/src/cmd.rs
