+++ opte-api/src/encap.rs
