+++ opte/Cargo.toml
@@ -11,7 +11,7 @@ engine = ["api"]
 # This feature indicates that OPTE is being built with std. This is
 # mostly useful to consumers of the API, providing convenient methods
 # for working with the API types in a std context.
-std = []
+std = ["opte-api/std"]
 
 [dependencies]
 cfg-if = "0.1"
@@ -24,6 +24,7 @@ cstr_core = "0.2.3"
 heapless = "0.7.9"
 illumos-ddi-dki = { path = "../illumos-ddi-dki" }
 illumos-sys-hdrs = { path = "../illumos-sys-hdrs" }
+opte-api = { path = "../opte-api", default-features = false }
 postcard = { version = "0.7.0", features = ["alloc"], default-features = false }
 #
 # XXX usdt is really a dev dependency, but when I move it there weird
